moved contents into master react app and tested successfully
