--- mud_game/src/actions/actionTypes.js
@@ -1,10 +1,0 @@
-export const REQUEST_START = "REQUEST_START";
-
-//register and login action types
-export const REGISTER_SUCCESS = "REGISTER_SUCCESS";
-export const REGISTER_FAILURE = "REGISTER_FAILURE";
-
-export const LOGIN_SUCCESS = "LOGIN_SUCCESS";
-export const LOGIN_FAILURE = "LOGIN_FAILURE";
-
-export const LOGOUT = "LOGOUT";
