--- mud_game/src/store/store.js
@@ -1,18 +1,0 @@
-import * as reducers from "../reducers/redurcers";
-import { combineReducers, createStore, compose, applyMiddleware } from "redux";
-import thunk from "redux-thunk";
-
-const monsterReducer = combineReducers({
-  onboard : reducers.onBoardingReducer
-});
-
-const store = createStore(
-  monsterReducer,
-  {},
-  compose(
-    applyMiddleware(thunk),
-    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-  )
-);
-
-export default store;
