+++ src/routes/index.js
@@ -6,12 +6,16 @@ import Layout from '../layout';
 
 import Home from '../components/home';
 import World from '../components/world';
+import FormikLoginForm from '../components/loginForm/LoginForm';
+import FormikRegisterForm from '../components/registerForm/RegisterForm'
 
 const AppRouter = () => {
 	return (
 		<BrowserRouter>
 			<Switch>
 				<Layout exact path="/" component={Home} />
+				<Layout path='/login' component={FormikLoginForm}/>
+				<Layout path='/register' component={FormikRegisterForm}/>
 				<Layout path="/game" component={World} />
 			</Switch>
 		</BrowserRouter>
