+++ src/state/store.js
@@ -3,8 +3,10 @@ import { applyMiddleware, compose, createStore, combineReducers } from 'redux';
 import thunk from 'redux-thunk';
 
 import playerReducer from './reducers/player';
+import authReducer from './reducers/auth';
 
 const rootReducer = combineReducers({
+	auth: authReducer,
 	player: playerReducer
 });
