+++ src/state/types/index.js
@@ -1,4 +1,16 @@
 /* eslint-disable import/prefer-default-export */
+
+export const REQUEST_START = "REQUEST_START";
+
+// Authentication
+export const REGISTER_SUCCESS = "REGISTER_SUCCESS";
+export const REGISTER_FAILURE = "REGISTER_FAILURE";
+
+export const LOGIN_SUCCESS = "LOGIN_SUCCESS";
+export const LOGIN_FAILURE = "LOGIN_FAILURE";
+
+export const LOGOUT = "LOGOUT";
+
 // Player Movement
 export const MOVE_PLAYER_UP = 'MOVE_PLAYER_UP';
 export const MOVE_PLAYER_DOWN = 'MOVE_PLAYER_DOWN';
